prima/route-planner/requirements.txt
flask==2.0.1
werkzeug==2.0.1
openpyxl==3.1.0
numpy==2.3.2
six
pyjwt
redis
geopy
google-cloud-pubsub
google-cloud-storage
google-cloud-core
memory-profiler
memoization
google-cloud-optimization
protobuf
scipy
shapely
here-location-services 